ASoC: amd: ps: add soundwire dma irq thread callback
Add acp pci driver Soundwire DMA irq thread callaback
for ACP6.3 platform.

Link: https://patch.msgid.link/20250207062819.1527184-7-Vijendar.Mukunda@amd.com
